Configures freestyle-logging module as crossproject
